Fixed and Cleaned up python implementations
